refactor: split commands.ts into focused modules (Closes #348)
Extract four new modules from the 1338-line commands.ts:
- config-panel-html.ts: HTML builder, nonce/escape helpers, platform defaults
- workspace-folder-resolver.ts: workspace folder resolution helpers
- debug-session-actions.ts: debug session start/restart/auto-start logic
- bundle-asset-installer.ts: bundled ROM asset install plan helpers

commands.ts shrinks to ~200 lines of pure command registration glue.
No behaviour changes; all 750 tests pass.
